+++ RealUI_Skins/RealUI_Skins.lua
@@ -42,12 +42,7 @@ local defaults = {
     }
 }
 
--- TODO: remove and convert to Aurora.Color
-RealUI.media = {
-    textures = {
-        plain = [[Interface\Buttons\WHITE8x8]],
-    },
-}
+RealUI.textures = private.textures
 
 local moddedFrames, pixelScale = {}
 local function ResetScale(frame)
@@ -170,6 +165,10 @@ function private.OnLoad()
 
     private.uiScale = private.skinsDB.uiModScale
     private.UpdateUIScale = RealUI.UpdateUIScale
+
+    for fontType, font in next, private.skinsDB.fonts do
+        private.font[fontType] = font.path or LSM:Fetch("font", font.name)
+    end
 
     local Base = Aurora.Base
     local Hook, Skin = Aurora.Hook, Aurora.Skin