+++ nibRealUI_Config/Advanced.lua
@@ -384,59 +384,15 @@ local core do
                     disabled = function() return not RealUI:GetModuleEnabled(MODNAME) end,
                     order = 40,
                     args = {
-                        opacity1 = {
-                            name = "Initial Dim",
-                            desc = "How dark to set the gameworld when you go AFK.",
-                            type = "range",
-                            isPercent = true,
-                            min = 0, max = 1, step = 0.05,
-                            get = function(info) return ScreenSaver.db.profile.general.opacity1 end,
-                            set = function(info, value) ScreenSaver.db.profile.general.opacity1 = value end,
-                            order = 10,
-                        },
-                        opacity2 = {
-                            name = "5min+ Dim",
-                            desc = "How dark to set the gameworld after 5 minutes of being AFK.",
-                            type = "range",
-                            isPercent = true,
-                            min = 0, max = 1, step = 0.05,
-                            get = function(info) return ScreenSaver.db.profile.general.opacity2 end,
-                            set = function(info, value) ScreenSaver.db.profile.general.opacity2 = value end,
-                            order = 20,
-                        },
                         combatwarning = {
                             name = "Combat Warning",
                             desc = "Play a warning sound if you enter combat while AFK.",
                             type = "toggle",
-                            get = function() return ScreenSaver.db.profile.general.combatwarning end,
+                            get = function() return ScreenSaver.db.profile.combatwarning end,
                             set = function(info, value)
-                                ScreenSaver.db.profile.general.combatwarning = value
+                                ScreenSaver.db.profile.combatwarning = value
                             end,
                             order = 30,
-                        },
-                    },
-                },
-                gap2 = {
-                    name = " ",
-                    type = "description",
-                    order = 41,
-                },
-                panel = {
-                    name = "Panel",
-                    type = "group",
-                    inline = true,
-                    disabled = function() return not RealUI:GetModuleEnabled(MODNAME) end,
-                    order = 50,
-                    args = {
-                        automove = {
-                            name = "Auto Move",
-                            desc = "Reposition the Panel up and down the screen once every minute.",
-                            type = "toggle",
-                            get = function() return ScreenSaver.db.profile.panel.automove end,
-                            set = function(info, value)
-                                ScreenSaver.db.profile.panel.automove = value
-                            end,
-                            order = 20,
                         },
                     },
                 },
@@ -447,33 +403,29 @@ local core do
         local MODNAME = "WorldMarker"
         local WorldMarker = RealUI:GetModule(MODNAME)
         worldMarker = {
-            name = "World Marker",
-            desc = "Quick access to World Markers.",
+            name = L["WorldMarker"],
+            desc = L["WorldMarkerDesc"],
             type = "group",
             childGroups = "tab",
             args = {
                 header = {
-                    name = "World Marker",
+                    name = L["WorldMarker"],
                     type = "header",
                     order = 10,
                 },
                 desc = {
-                    name = "Quick access to World Markers.",
+                    name = L["WorldMarkerDesc"],
                     type = "description",
                     fontSize = "medium",
                     order = 20,
                 },
                 enabled = {
-                    name = "Enabled",
-                    desc = "Enable/Disable the WorldMarker module.",
+                    name = L["General_Enabled"],
+                    desc = L["General_EnabledDesc"]:format(L[MODNAME]),
                     type = "toggle",
                     get = function() return RealUI:GetModuleEnabled(MODNAME) end,
                     set = function(info, value)
-                        if not _G.InCombatLockdown() then
-                            RealUI:SetModuleEnabled(MODNAME, value)
-                        else
-                            _G.print("|cff0099ffRealUI: |r World Marker can't be enabled or disabled during combat.")
-                        end
+                        RealUI:SetModuleEnabled(MODNAME, value)
                     end,
                     order = 30,
                 },
@@ -483,7 +435,7 @@ local core do
                     order = 31,
                 },
                 visibility = {
-                    name = "Show the World Marker in..",
+                    name = L["WorldMarker_Show"],
                     type = "group",
                     disabled = function() return not RealUI:GetModuleEnabled(MODNAME) end,
                     order = 40,
@@ -527,6 +479,16 @@ local core do
                                 WorldMarker:UpdateVisibility()
                             end,
                             order = 40,
+                        },
+                        none = {
+                            name = _G.WORLD,
+                            type = "toggle",
+                            get = function(info) return WorldMarker.db.profile.visibility.none end,
+                            set = function(info, value)
+                                WorldMarker.db.profile.visibility.none = value
+                                WorldMarker:UpdateVisibility()
+                            end,
+                            order = 50,
                         },
                     },
                 },